% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPDfunctions.R
\name{addActivity}
\alias{addActivity}
\title{addActivity function}
\usage{
addActivity(Activity, updatedList)
}
\arguments{
\item{Activity}{file name for activity table.}

\item{supertabDrogue}{file name for any table containing drogue codes.}
}
\value{
A new table.
}
\description{
High level function. Add two new columns to drogue code table.
             Both columns come from "Cecile's table"
}
\examples{

actTable <-addActivity ("activity2.txt", "supertabDrogue.csv")
actTable <-addActivity ("activity2.txt", DFDrogue)

}
