% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPDfunctions.R
\name{addCategory}
\alias{addCategory}
\title{addCategory function}
\usage{
addCategory(superTabroute, n = 6)
}
\arguments{
\item{kfile}{file name for route table.}
}
\value{
New table.
}
\description{
High level function. Add new (Health Canada Category) column to route table
}
\examples{
 updatedTable<-addCategory("superTab2.csv")
 updatedTable<-addCategory(tab)
}
