% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPDfunctions.R
\name{addExtraVetcols}
\alias{addExtraVetcols}
\title{addExtraVetcols function}
\usage{
addExtraVetcols(extravetcol, supertabVet)
}
\arguments{
\item{extravetcol}{file name for extra vet columns.}

\item{supertabVet}{file name for any table containing drogue codes and vet vars.}
}
\value{
A new table.
}
\description{
High level function. Add extra vet columns
}
\examples{

actTable <-aaddExtraVetcols ("activity2.txt", "supertabDrogue.csv")
actTable <-addExtraVetcols ("activity2.txt", DFDrogue)

}
