% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPDfunctions.R
\name{AddMergedIngredNames}
\alias{AddMergedIngredNames}
\title{AddMergedIngredNames function.}
\usage{
AddMergedIngredNames(ingredfile)
}
\arguments{
\item{ingredfile}{name}
}
\description{
High level function.
              Add new (Health Canada Category) column  for  merged ingred names
}
\examples{

updatedTab<-AddMergedIngredNames("tabIged.csv")

}
