% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPDfunctions.R
\name{AddMergedIngredNames2}
\alias{AddMergedIngredNames2}
\title{AddMergedIngredNames2 function.}
\usage{
AddMergedIngredNames2(ingredfile)
}
\arguments{
\item{ingredfile}{name}
}
\description{
High level function for  transposed ingredient table
              Add new (Health Canada Category) column  for  merged ingred names
}
\examples{

updatedTab<-AddMergedIngredNames2("tabIged.csv")
updatedTab<-AddMergedIngredNames2(tabIged)

}
