% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPDextractor.R
\name{biosimilar}
\alias{biosimilar}
\title{biosimilar class}
\description{
R6 Class representing biosimilar drugs.
}
\details{
The biosimilar class has 2  methods in addition to 6 other methods and a field that inherits from the
parent class (DPDfile).
}
\examples{

## ------------------------------------------------
## Method `biosimilar$new`
## ------------------------------------------------

obj<-biosimilar$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/biosimilar.txt")


## ------------------------------------------------
## Method `biosimilar$select`
## ------------------------------------------------

obj<-biosimilar$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/biosimilar.txt")
obj$select()

}
\section{Super class}{
\code{\link[DPDextractor3:DPDfile]{DPDextractor3::DPDfile}} -> \code{biosimilar}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-biosimilar-new}{\code{biosimilar$new()}}
\item \href{#method-biosimilar-select}{\code{biosimilar$select()}}
\item \href{#method-biosimilar-clone}{\code{biosimilar$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DPDextractor3" data-topic="DPDfile" data-id="field_search"><a href='../../DPDextractor3/html/DPDfile.html#method-DPDfile-field_search'><code>DPDextractor3::DPDfile$field_search()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-biosimilar-new"></a>}}
\if{latex}{\out{\hypertarget{method-biosimilar-new}{}}}
\subsection{Method \code{new()}}{
Create a new biosimilar object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{biosimilar$new(fname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fname}}{file and path names.}

\item{\code{dpdlabel}}{distinctive element of the ther name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new ther object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-biosimilar$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/biosimilar.txt")

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-biosimilar-select"></a>}}
\if{latex}{\out{\hypertarget{method-biosimilar-select}{}}}
\subsection{Method \code{select()}}{
Chooses the table columns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{biosimilar$select(columns = c("DRUG_CODE", "BIOSIMILAR"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{columns}}{field names.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-biosimilar$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/biosimilar.txt")
obj$select()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-biosimilar-clone"></a>}}
\if{latex}{\out{\hypertarget{method-biosimilar-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{biosimilar$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
