% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPDfunctions.R
\name{CFmergeFiles}
\alias{CFmergeFiles}
\title{CFmergeFiles}
\usage{
CFmergeFiles(
  b = c(1, 0, 0, 1),
  dpdlabel = c("drug", "ingred", "route", "form", "status", "comp", "package", "vet"),
  filt = TRUE,
  fout = "superTabIngDrugrouteformCompStaVetPack_CF6.csv"
)
}
\arguments{
\item{b}{boolean vector to select whether or not to transpose the ingred, form, route and vet tables}

\item{dpdlabel}{vector indicating the order in which the files are merged.}

\item{path}{to the directory where the DPD files are located}
}
\value{
A merged table.
}
\description{
function for vertical and horizontal merging of dpd files in the order required
to generate Cecile and Nicolas' excel file or any other similar file. Drug and ingred must be the first two labels
}
\examples{

CFmergeFiles()
CFmergeFiles(b=c(1,0,1,1),dpdlabel=c("drug","ingred", "route", "form", "status","comp", "package", "vet"), filt=TRUE, fout="superTabIngDrugrouteformCompStaVetPack_CF2.csv")
}
