% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPDfunctions.R
\name{CFmergeFiles2}
\alias{CFmergeFiles2}
\title{CFmergeFiles2}
\usage{
CFmergeFiles2(
  b = c(1, 0, 0, 1),
  dpdlabel = c("drug", "ingred", "form", "route", "status", "comp", "package", "vet"),
  filt = TRUE,
  fout = "superTabIngDrugrouteformCompStaVetPack_CF0.csv",
  bm1 = 1,
  bm2 = 1,
  extrac = TRUE
)
}
\arguments{
\item{b}{boolean vector to select whether or not to transpose the ingred, form, route and vet tables}

\item{dpdlabel}{vector indicating the order in which the files are merged.}

\item{bm1}{this parameter allows access to the argument (0: all=T, 1: all.x=T, 2: all.y=T) of the merge  function for the table Form}

\item{bm2}{this parameter allows access to the argument (0: all=T, 1: all.x=T, 2: all.y=T)  of the merge  function for the table Route}

\item{extrac}{Boolean parameters to add extra columns to vet subtable}

\item{path}{to the directory where the DPD files are located}
}
\value{
A merged table.
}
\description{
function for vertical and horizontal merging of dpd files in the order required
to generate a file similar to the Cecile and Nicolas' excel file. Drug and ingred must be the first two labels
}
\examples{

CFmergeFiles2() #sintaxis to generate a file very similar to the "Cecile and Nicolas' excel file"
}
