% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPDextractor.R
\name{comp}
\alias{comp}
\title{comp class}
\description{
R6 Class representing drug manufacturers.
}
\details{
#' The comp class has 2  methods in addition to 6 other methods and a field that inherits from the
parent class (DPDfile).
Every drug has one and only one manufacturer. However, one manufacturer may be linked to several drugs.
 the comp table has 18 fields:"DRUG_CODE", "MFR_CODE", "COMPANY_CODE", "COMPANY_NAME", "COMPANY_TYPE",
 "ADDRESS_MAILING_FLAG", "ADDRESS_BILLING_FLAG", "ADDRESS_NOTIFICATION_FLAG", "ADDRESS_OTHER", "SUITE_NUMBER",
 "STREET_NAME","CITY_NAME", "PROVINCE", "COUNTRY", "POSTAL_CODE", "POST_OFFICE_BOX",
 "PROVINCE_F", "COUNTRY_F"
}
\examples{

## ------------------------------------------------
## Method `comp$new`
## ------------------------------------------------

obj<-comp$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/comp.txt")
obj<-comp$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/comp_ia.txt")


## ------------------------------------------------
## Method `comp$select`
## ------------------------------------------------

obj<-drug$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/comp.txt")
obj$select()
}
\section{Super class}{
\code{\link[DPDextractor3:DPDfile]{DPDextractor3::DPDfile}} -> \code{comp}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-comp-new}{\code{comp$new()}}
\item \href{#method-comp-select}{\code{comp$select()}}
\item \href{#method-comp-clone}{\code{comp$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DPDextractor3" data-topic="DPDfile" data-id="field_search"><a href='../../DPDextractor3/html/DPDfile.html#method-DPDfile-field_search'><code>DPDextractor3::DPDfile$field_search()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-comp-new"></a>}}
\if{latex}{\out{\hypertarget{method-comp-new}{}}}
\subsection{Method \code{new()}}{
Create a new comp object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{comp$new(fname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fname}}{file and path names.}

\item{\code{dpdlabel}}{distinctive element of the comp name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new comp object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-comp$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/comp.txt")
obj<-comp$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/comp_ia.txt")

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-comp-select"></a>}}
\if{latex}{\out{\hypertarget{method-comp-select}{}}}
\subsection{Method \code{select()}}{
Chooses the table columns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{comp$select(columns = c("DRUG_CODE", "COMPANY_CODE", "COMPANY_NAME"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{columns}}{field names.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-drug$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/comp.txt")
obj$select()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-comp-clone"></a>}}
\if{latex}{\out{\hypertarget{method-comp-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{comp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
