% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPDfunctions.R
\name{datasetVmerge}
\alias{datasetVmerge}
\title{datasetVmerge function}
\usage{
datasetVmerge(
  txt_files,
  dpdlabel,
  f = "listAntibiodine.txt",
  stat = c(1, 1, 1, 1),
  filt = TRUE
)
}
\arguments{
\item{txt_files}{file name.}

\item{dpdlabel}{dintinctive component name: status", "drug","route", "ingred", "comp", "package" OR "vet"}

\item{f}{list of antibiotics and A families}

\item{stat}{boolean vector to select the four drug status: marketed, approved, cancelled and dormant}

\item{filt}{boolean parameter to choose antibiotic filtering (filt=TRUE)}
}
\value{
A merged table.
}
\description{
Function for vertical merging of tables containing drugs of different statuses
}
\examples{

sta<-datasetVmerge(txt_files, dpdlabel="status")
ing<-datasetVmerge(txt_files, dpdlabel="ingred")
ing<-datasetVmerge(txt_files, dpdlabel="ingred", stat=c(1, 1, 0, 0))              # MEANING: MARKETED and APPROVED ANTIBIOTICS
ing<-datasetVmerge(txt_files, dpdlabel="ingred", stat=c(1, 0, 0, 0), filt=FALSE)  # MEANING: MARKETED DRUGS
f="listAntibiodine.txt"
}
