% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPDfunctions.R
\name{datasetVTransmerge}
\alias{datasetVTransmerge}
\title{datasetVTransmerge}
\usage{
datasetVTransmerge(
  txt_files,
  dpdlabel,
  f = "listAntibiodine.txt",
  stat = c(1, 1, 1, 1),
  filt = TRUE,
  spe = TRUE
)
}
\arguments{
\item{txt_files}{file name.}

\item{dpdlabel}{dintinctive component name: "status", "drug","route", "ingred", "comp", "package" OR "vet"}

\item{f}{list of antibiotics and A families}

\item{stat}{boolean vector to select the four drug status: marketed, approved, cancelled and dormant}

\item{filt}{boolean parameter to choose antibiotic filtering (filt=TRUE)}

\item{TR}{boolean parameter to choose transposition for ingred, vet, form and route tasbles}
}
\value{
A merged table.
}
\description{
Function for vertical merging of tables containing drugs of different statuses and
selective transposition for "vet" and "ingred"
}
\examples{
sta<-datasetVTransmerge(txt_files, dpdlabel="status")
ing<-datasetVTransmerge(txt_files, dpdlabel="ingred")
ing<-datasetVTransmerge(txt_files, dpdlabel="ingred", stat=c(1, 1, 0, 0))              # MEANING: MARKETED and APPROVED ANTIBIOTICS
ing<-datasetVTransmerge(txt_files, dpdlabel="ingred", stat=c(1, 0, 0, 0), filt=FALSE)  # MEANING: MARKETED DRUGS
 f="listAntibiodine.txt"
}
