% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPDextractor.R
\name{DPDfile}
\alias{DPDfile}
\title{DPDfile class}
\description{
R6 Class representing a DPD file.The DPDfile class has 5 methods and 1 private attribute, the dataframe (dfm) with the DPD file content
}
\details{
The 5 methods of this class are header, select,field_search, getter/setter functions and constructor.
DPDfile is the supoerclass of all protect classes
}
\examples{

## ------------------------------------------------
## Method `DPDfile$new`
## ------------------------------------------------

obj<-DPDfile$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/vet.txt")
Table<-obj$DPDtable
obj$DPDtable<-Table
obj<-DPDfile$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/vet.txt")


## ------------------------------------------------
## Method `DPDfile$select`
## ------------------------------------------------

obj1$select()

## ------------------------------------------------
## Method `DPDfile$field_search`
## ------------------------------------------------

obj$field_search ("STATUS") #for string searches
          bj$field_search ("STATUS", b=TRUE) #for sboolean searches

        If b==TRUE, boolean value representing absence/presence of a name or
        prefix/infix/suffix in the set of field names.
        If b==FALSE,Name(s) of the table(s) where the field is located

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dfm}}{table or data frame derived from the corresponding DPD file.                     #'}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{dfm}}{table or data frame derived from the corresponding DPD file.                     #'}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DPDfile-new}{\code{DPDfile$new()}}
\item \href{#method-DPDfile-select}{\code{DPDfile$select()}}
\item \href{#method-DPDfile-field_search}{\code{DPDfile$field_search()}}
\item \href{#method-DPDfile-clone}{\code{DPDfile$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DPDfile-new"></a>}}
\if{latex}{\out{\hypertarget{method-DPDfile-new}{}}}
\subsection{Method \code{new()}}{
active binding for dpd tables


Getter/Setter method for DPDfile object


Create a new DPDfile class object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DPDfile$new(fname = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fname}}{file and path names.}

\item{\code{dpdlabel}}{distinctive elements of DPD file names.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `DPDfile` object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-DPDfile$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/vet.txt")
Table<-obj$DPDtable
obj$DPDtable<-Table
}
\if{html}{\out{</div>}}

\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-DPDfile$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/vet.txt")

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DPDfile-select"></a>}}
\if{latex}{\out{\hypertarget{method-DPDfile-select}{}}}
\subsection{Method \code{select()}}{
Choose the subset of the table columns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DPDfile$select(columns)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{columns}}{column names.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj1$select()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DPDfile-field_search"></a>}}
\if{latex}{\out{\hypertarget{method-DPDfile-field_search}{}}}
\subsection{Method \code{field_search()}}{
this function searches for fields that have a given name or substring
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DPDfile$field_search(field, b = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field}}{field name or prefix/suffix.}

\item{\code{b}}{Boolean parameter to define the returned values .}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
string containing found fields or boolean value
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj$field_search ("STATUS") #for string searches
          bj$field_search ("STATUS", b=TRUE) #for sboolean searches

        If b==TRUE, boolean value representing absence/presence of a name or
        prefix/infix/suffix in the set of field names.
        If b==FALSE,Name(s) of the table(s) where the field is located

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DPDfile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DPDfile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DPDfile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
