% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPDextractor.R
\name{drug}
\alias{drug}
\title{drug class}
\description{
R6 Class representing drug products.
}
\details{
The drug class has 3  methods in addition to 6 other methods and a field that inherits from the
parent class (DPDfile).The drug table has 14 fields:  "DRUG_CODE", "PRODUCT_CATEGORIZATION", "CLASS", "DRUG_IDENTIFICATION_NUMBER",
"BRAND_NAME", "DESCRIPTOR", "PEDIATRIC_FLAG","ACCESSION_NUMBER","NUMBER_OF_AIS", "LAST_UPDATE_DATE",
"AI_GROUP_NO","CLASS_F","BRAND_NAME_F", "DESCRIPTOR_F"
}
\examples{

## ------------------------------------------------
## Method `drug$new`
## ------------------------------------------------

obj<-drug$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/drug.txt")
obj<-drug$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/drug_ia.txt")


## ------------------------------------------------
## Method `drug$select`
## ------------------------------------------------

obj<-drug$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/drug.txt")
obj$select()

## ------------------------------------------------
## Method `drug$Drug_search`
## ------------------------------------------------

obj<-drug$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/drug.txt")
obj$select()

}
\section{Super class}{
\code{\link[DPDextractor3:DPDfile]{DPDextractor3::DPDfile}} -> \code{drug}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-drug-new}{\code{drug$new()}}
\item \href{#method-drug-select}{\code{drug$select()}}
\item \href{#method-drug-Drug_search}{\code{drug$Drug_search()}}
\item \href{#method-drug-clone}{\code{drug$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DPDextractor3" data-topic="DPDfile" data-id="field_search"><a href='../../DPDextractor3/html/DPDfile.html#method-DPDfile-field_search'><code>DPDextractor3::DPDfile$field_search()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-drug-new"></a>}}
\if{latex}{\out{\hypertarget{method-drug-new}{}}}
\subsection{Method \code{new()}}{
Create a new drug object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{drug$new(fname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fname}}{file and path names.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new drug object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-drug$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/drug.txt")
obj<-drug$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/drug_ia.txt")

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-drug-select"></a>}}
\if{latex}{\out{\hypertarget{method-drug-select}{}}}
\subsection{Method \code{select()}}{
Chooses the table columns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{drug$select(
  columns = c("DRUG_CODE", "CLASS", "DRUG_IDENTIFICATION_NUMBER", "BRAND_NAME",
    "NUMBER_OF_AIS", "AI_GROUP_NO")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{columns}}{field names.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-drug$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/drug.txt")
obj$select()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-drug-Drug_search"></a>}}
\if{latex}{\out{\hypertarget{method-drug-Drug_search}{}}}
\subsection{Method \code{Drug_search()}}{
Chooses the table columns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{drug$Drug_search(
  drugCode,
  fields = c("DRUG_CODE", "CLASS", "DRUG_IDENTIFICATION_NUMBER", "BRAND_NAME",
    "NUMBER_OF_AIS")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{drugCode}}{code}

\item{\code{fields}}{drug table fields}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-drug$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/drug.txt")
obj$select()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-drug-clone"></a>}}
\if{latex}{\out{\hypertarget{method-drug-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{drug$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
