% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPDfunctions.R
\name{FileMerge}
\alias{FileMerge}
\title{FileMerge}
\usage{
FileMerge(
  txt_files,
  b = c(1, 0, 0, 1),
  dpdlabel = c("drug", "status", "route", "form", "comp", "ingred", "package", "vet"),
  filt = TRUE,
  outf = TRUE
)
}
\arguments{
\item{txt_files}{set of file names.}

\item{b}{boolean vector to select whether or not to transpose the ingred, form, route and vet tables}

\item{dpdlabel}{vector indicating the files to be merged.}

\item{path}{to the directory where the DPD files are located}
}
\value{
Vmerged tables in DPD files
}
\description{
function for merging pd files and generating new  and separate
csv files
}
\examples{

txt_files <- list.files(path = getwd(), pattern = "txt")
FileMerge(txt_files)
FileMerge(txt_files, b=c(0,0,0,0))
FileMerge(txt_files, b=c(1,1,0,1))
FileMerge(txt_files, b=c(1,1,1,1))
FileMerge(txt_files, b=c(0,0,1,1),  dpdlabel=c("ingred", "form", "route","vet"))
}
