% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPDextractor.R
\name{form}
\alias{form}
\title{form class}
\description{
R6 Class representing drug forms.
}
\details{
The package class has 2  methods in addition to 6 other methods and a field that inherits from the
parent class (DPDfile).
There are several pharmaceutical forms in this table: SOLUTION, TABLET,SPRAY, WIPE, LOTION,
 OINTMENT, POWDER, CAPSULE and CREAM.
The same product can have several different forms
the package table has 4 fields:"DRUG_CODE", "PHARM_FORM_CODE", "PHARMACEUTICAL_FORM",
"PHARMACEUTICAL_FORM_F"
}
\examples{

## ------------------------------------------------
## Method `form$new`
## ------------------------------------------------

obj<-form$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/form.txt")
obj<-form$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/form_ia.txt")


## ------------------------------------------------
## Method `form$select`
## ------------------------------------------------

obj<-form$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/form.txt")
obj$select()


## ------------------------------------------------
## Method `form$CTranspose`
## ------------------------------------------------

obj<-form$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/form.txt")
obj$CTranspose()
}
\section{Super class}{
\code{\link[DPDextractor3:DPDfile]{DPDextractor3::DPDfile}} -> \code{form}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-form-new}{\code{form$new()}}
\item \href{#method-form-select}{\code{form$select()}}
\item \href{#method-form-CTranspose}{\code{form$CTranspose()}}
\item \href{#method-form-clone}{\code{form$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DPDextractor3" data-topic="DPDfile" data-id="field_search"><a href='../../DPDextractor3/html/DPDfile.html#method-DPDfile-field_search'><code>DPDextractor3::DPDfile$field_search()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-form-new"></a>}}
\if{latex}{\out{\hypertarget{method-form-new}{}}}
\subsection{Method \code{new()}}{
Create a new form object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{form$new(fname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fname}}{file and path names.}

\item{\code{dpdlabel}}{distinctive element of the form name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new form object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-form$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/form.txt")
obj<-form$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/form_ia.txt")

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-form-select"></a>}}
\if{latex}{\out{\hypertarget{method-form-select}{}}}
\subsection{Method \code{select()}}{
Chooses the table columns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{form$select(columns = c("DRUG_CODE", "PHARMACEUTICAL_FORM"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{columns}}{field names.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-form$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/form.txt")
obj$select()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-form-CTranspose"></a>}}
\if{latex}{\out{\hypertarget{method-form-CTranspose}{}}}
\subsection{Method \code{CTranspose()}}{
Selective transposition for form table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{form$CTranspose()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-form$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/form.txt")
obj$CTranspose()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-form-clone"></a>}}
\if{latex}{\out{\hypertarget{method-form-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{form$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
