% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPDfunctions.R
\name{HVmerge}
\alias{HVmerge}
\title{HVmerge}
\usage{
HVmerge(
  txt_files,
  b = TRUE,
  dpdlabel = c("status", "drug", "route", "form", "ingred", "comp", "package", "ther",
    "vet"),
  f = "listAntibiodine.txt"
)
}
\arguments{
\item{txt_files}{set of file names.}

\item{b}{boolean param to choose whether to transpose or not}

\item{dpdlabel}{vector indicating the order in which the files are merged.}
}
\value{
A merged table.
}
\description{
Function for horizontal and vertical merging in a given ordering of previously chosen tables
containing drugs of different statuses.
}
\examples{
txt_files <- list.files(path = "D:/WIAAgrosante/Mapaq-Antibioticos/DPD", pattern = "txt", full.names=TRUE)
df<-HVmerge(txt_files)
df<-HVmerge(txt_files, b=FALSE)
df<-HVmerge(txt_files, b=TRUE,  dpdlabel=c("drug", "ingred", "vet", "package"))
}
