% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPDextractor.R
\name{ingred}
\alias{ingred}
\title{ingred class}
\description{
R6 Class representing the active ingredients of a given drug product
}
\details{
The ingred class has 8 methods in addition to 6 other methods and a field that inherits from the
parent class (DPDfile).A drug may have one or more active ingredients.
}
\examples{

## ------------------------------------------------
## Method `ingred$new`
## ------------------------------------------------

obj<-ingred$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/ingred.txt")
obj<-ingred$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/ingred_ap.txt")


## ------------------------------------------------
## Method `ingred$select`
## ------------------------------------------------

obj<-ingred$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/ingred_ia.txt")
obj$select(c("DRUG_CODE", "INGREDIENT"))
obj$select()

## ------------------------------------------------
## Method `ingred$CTranspose`
## ------------------------------------------------

obj<-ingred$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/ingred.txt")
obj$CTranspose()

## ------------------------------------------------
## Method `ingred$freqDrugCode`
## ------------------------------------------------

obj<-ingred$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/ingred.txt")
dc<-obj$DPDtable[,"DRUG_CODE"]
f<-obj$freqDrugCode(dc)

## ------------------------------------------------
## Method `ingred$filtering`
## ------------------------------------------------

obj<-ingred$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/ingred.txt")
df<-obj$filtering(a, b, c, d)


## ------------------------------------------------
## Method `ingred$Afiltering`
## ------------------------------------------------

obj<-ingred$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/ingred.txt")
obj$Afiltering(ab, af)
obj$Afiltering(ab, af, b=TRUE, verbose=TRUE)

## ------------------------------------------------
## Method `ingred$ACfiltering`
## ------------------------------------------------

obj<-ingred$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/ingred.txt")
obj$ACfiltering(ab, af, ac)
obj$ACfiltering(ab, af, ac, b=TRUE, verbose=TRUE)

}
\section{Super class}{
\code{\link[DPDextractor3:DPDfile]{DPDextractor3::DPDfile}} -> \code{ingred}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ingred-new}{\code{ingred$new()}}
\item \href{#method-ingred-select}{\code{ingred$select()}}
\item \href{#method-ingred-CTranspose}{\code{ingred$CTranspose()}}
\item \href{#method-ingred-freqDrugCode}{\code{ingred$freqDrugCode()}}
\item \href{#method-ingred-freqIngred}{\code{ingred$freqIngred()}}
\item \href{#method-ingred-filtering}{\code{ingred$filtering()}}
\item \href{#method-ingred-Afiltering}{\code{ingred$Afiltering()}}
\item \href{#method-ingred-ACfiltering}{\code{ingred$ACfiltering()}}
\item \href{#method-ingred-clone}{\code{ingred$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DPDextractor3" data-topic="DPDfile" data-id="field_search"><a href='../../DPDextractor3/html/DPDfile.html#method-DPDfile-field_search'><code>DPDextractor3::DPDfile$field_search()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ingred-new"></a>}}
\if{latex}{\out{\hypertarget{method-ingred-new}{}}}
\subsection{Method \code{new()}}{
Create a new ingred object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ingred$new(fname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fname}}{file and path names.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new ingred object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-ingred$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/ingred.txt")
obj<-ingred$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/ingred_ap.txt")

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ingred-select"></a>}}
\if{latex}{\out{\hypertarget{method-ingred-select}{}}}
\subsection{Method \code{select()}}{
choose the ingred fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ingred$select(
  columns = c("DRUG_CODE", "BASE", "INGREDIENT", "STRENGTH", "STRENGTH_UNIT",
    "STRENGTH_TYPE")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{columns}}{a vector of route field names}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-ingred$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/ingred_ia.txt")
obj$select(c("DRUG_CODE", "INGREDIENT"))
obj$select()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ingred-CTranspose"></a>}}
\if{latex}{\out{\hypertarget{method-ingred-CTranspose}{}}}
\subsection{Method \code{CTranspose()}}{
Selective transposition of drugs with at least one active ingredient.
based on maximum number of active ingredients
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ingred$CTranspose(a = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{indicating the relevance of antib family}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-ingred$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/ingred.txt")
obj$CTranspose()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ingred-freqDrugCode"></a>}}
\if{latex}{\out{\hypertarget{method-ingred-freqDrugCode}{}}}
\subsection{Method \code{freqDrugCode()}}{
compute freq distribution of the number of active ingredients
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ingred$freqDrugCode(dc)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dc}}{drug code vector}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
table of frequency of the nomber of active ingredients per drug
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-ingred$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/ingred.txt")
dc<-obj$DPDtable[,"DRUG_CODE"]
f<-obj$freqDrugCode(dc)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ingred-freqIngred"></a>}}
\if{latex}{\out{\hypertarget{method-ingred-freqIngred}{}}}
\subsection{Method \code{freqIngred()}}{
compute freq distribution of active ingredients
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ingred$freqIngred(ingred)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dc}}{drug code vector}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
table of frequency of active ingredients
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ingred-filtering"></a>}}
\if{latex}{\out{\hypertarget{method-ingred-filtering}{}}}
\subsection{Method \code{filtering()}}{
select active ingredients that are antibiotics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ingred$filtering(
  Antib_list,
  AntiFam_list,
  SCcat_list,
  antibacterial,
  antiprotozoal
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Antib_list}}{list of antibiotics.}

\item{\code{AntiFam_list}}{list of antibiotic families}

\item{\code{SCcat_list}}{list of Canada Health categories}

\item{\code{antibacterial}}{Boolean variable indicating the presence/absence of antibacterial activity.}

\item{\code{antiprotozoal}}{Boolean variable indicating the presence/absence of antiprotozoal activity.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
table with four columns of the active ingredients that are antibiotics:
 name of the active ingredient, name of the corresponding antibiotic family,
 name of the Canada Health category, as well as the antibacterial or antiprotozoal activity
 of the corresponding active ingredient
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-ingred$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/ingred.txt")
df<-obj$filtering(a, b, c, d)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ingred-Afiltering"></a>}}
\if{latex}{\out{\hypertarget{method-ingred-Afiltering}{}}}
\subsection{Method \code{Afiltering()}}{
This function selects the active ingredients that are antibiotics and prints their families.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ingred$Afiltering(Antib_list, AntiFam_list, b = TRUE, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Antib_list}}{list of antibiotics.}

\item{\code{AntiFam_list}}{list of antibiotic families}

\item{\code{b}}{boolean parameter representing the method used to select the antibiotics. Recommended default value:TRUE}

\item{\code{verbose}}{boolean parameter for the complexity of the name of active ingredients. Full name:verbose=TRUE}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
table of five columns for the active ingredients that are antibiotics:
name of active ingredient, drug code, dose, dose units and antibiotic family.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-ingred$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/ingred.txt")
obj$Afiltering(ab, af)
obj$Afiltering(ab, af, b=TRUE, verbose=TRUE)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ingred-ACfiltering"></a>}}
\if{latex}{\out{\hypertarget{method-ingred-ACfiltering}{}}}
\subsection{Method \code{ACfiltering()}}{
This function selects the active ingredients that are antibiotics, prints their health category as well as the antibiotic family.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ingred$ACfiltering(
  Antib_list,
  AntiFam_list,
  AntiCatSC_list,
  b = TRUE,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Antib_list}}{list of antibiotics.}

\item{\code{AntiFam_list}}{list of antibiotic families}

\item{\code{b}}{boolean parameter representing the method used to select the antibiotics. Recommended default value:TRUE}

\item{\code{verbose}}{boolean parameter for the complexity of the name of active ingredients. Full name:verbose=TRUE}

\item{\code{AntiCatSC_lis}}{list of Canada health categories}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
table of 6 columns for the active ingredients that are antibiotics:
name of active ingredient, drug code, dose, dose units,  antibiotic family and Canada health category
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-ingred$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/ingred.txt")
obj$ACfiltering(ab, af, ac)
obj$ACfiltering(ab, af, ac, b=TRUE, verbose=TRUE)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ingred-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ingred-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ingred$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
