% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPDfunctions.R
\name{kdataset}
\alias{kdataset}
\title{kdataset function}
\usage{
kdataset(kfile, dpdlabel)
}
\arguments{
\item{kfile}{file name.}

\item{dpdlabel}{dintinctive component name.}
}
\value{
A new selected object.
}
\description{
High level function. Object selector
}
\examples{
obj1<-kdataset(txt_files[k], "ingred")

}
