% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPDextractor.R
\name{pharm}
\alias{pharm}
\title{pharm class}
\description{
R6 Class representing pharmaceutical STD.
}
\details{
The package class has 2  methods in addition to 6 other methods and a field that inherits from the
parent class (DPDfile).
the pharm table has 4 fields:"DRUG_CODE" and  "PHARMACEUTICAL_STD"
}
\examples{

## ------------------------------------------------
## Method `pharm$new`
## ------------------------------------------------

obj<-pharm$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/pharm.txt")
obj<-pharm$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/pharm_ia.txt")


## ------------------------------------------------
## Method `pharm$select`
## ------------------------------------------------

obj<-pharm$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/pharm.txt")
obj$select()

}
\section{Super class}{
\code{\link[DPDextractor3:DPDfile]{DPDextractor3::DPDfile}} -> \code{pharm}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-pharm-new}{\code{pharm$new()}}
\item \href{#method-pharm-select}{\code{pharm$select()}}
\item \href{#method-pharm-clone}{\code{pharm$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DPDextractor3" data-topic="DPDfile" data-id="field_search"><a href='../../DPDextractor3/html/DPDfile.html#method-DPDfile-field_search'><code>DPDextractor3::DPDfile$field_search()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pharm-new"></a>}}
\if{latex}{\out{\hypertarget{method-pharm-new}{}}}
\subsection{Method \code{new()}}{
Create a new pharm object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pharm$new(fname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fname}}{file and path names.}

\item{\code{dpdlabel}}{distinctive element of the pharm name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new pharm object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-pharm$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/pharm.txt")
obj<-pharm$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/pharm_ia.txt")

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pharm-select"></a>}}
\if{latex}{\out{\hypertarget{method-pharm-select}{}}}
\subsection{Method \code{select()}}{
Chooses the table columns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pharm$select(columns = c("DRUG_CODE", "PHARMACEUTICAL_STD"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{columns}}{field names.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-pharm$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/pharm.txt")
obj$select()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pharm-clone"></a>}}
\if{latex}{\out{\hypertarget{method-pharm-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pharm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
