% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPDextractor.R
\name{route}
\alias{route}
\title{route class}
\description{
R6 Class representing the DRUG product administration routes
}
\details{
The route class has 5 methods in addition to 6 other methods and a field that inherits from the
parent class (DPDfile).A drug may have one or more routes of administration.
}
\examples{

## ------------------------------------------------
## Method `route$new`
## ------------------------------------------------

obj<-route$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/route.txt")
obj<-route$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/route_ap.txt")

## ------------------------------------------------
## Method `route$select`
## ------------------------------------------------

obj<-route$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/route.txt", "route")
obj$select()
obj$select(c("DRUG_CODE", "ROUTE_OF_ADMINISTRATION_CODE", "ROUTE_OF_ADMINISTRATION"))

## ------------------------------------------------
## Method `route$CTranspose`
## ------------------------------------------------

obj<-route$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/route.txt", "route")
obj$select()
obj$CTranspose()
}
\section{Super class}{
\code{\link[DPDextractor3:DPDfile]{DPDextractor3::DPDfile}} -> \code{route}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-route-new}{\code{route$new()}}
\item \href{#method-route-select}{\code{route$select()}}
\item \href{#method-route-CTranspose}{\code{route$CTranspose()}}
\item \href{#method-route-clone}{\code{route$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DPDextractor3" data-topic="DPDfile" data-id="field_search"><a href='../../DPDextractor3/html/DPDfile.html#method-DPDfile-field_search'><code>DPDextractor3::DPDfile$field_search()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-route-new"></a>}}
\if{latex}{\out{\hypertarget{method-route-new}{}}}
\subsection{Method \code{new()}}{
Create a new route object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{route$new(fname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fname}}{file and path names.}

\item{\code{dpdlabel}}{distinctive element of the route name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new route object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-route$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/route.txt")
obj<-route$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/route_ap.txt")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-route-select"></a>}}
\if{latex}{\out{\hypertarget{method-route-select}{}}}
\subsection{Method \code{select()}}{
choose the route fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{route$select(columns = c("DRUG_CODE", "ROUTE_OF_ADMINISTRATION"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{columns}}{a vector of route field names}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-route$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/route.txt", "route")
obj$select()
obj$select(c("DRUG_CODE", "ROUTE_OF_ADMINISTRATION_CODE", "ROUTE_OF_ADMINISTRATION"))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-route-CTranspose"></a>}}
\if{latex}{\out{\hypertarget{method-route-CTranspose}{}}}
\subsection{Method \code{CTranspose()}}{
Selective transposition of drugs with at least one route of administration based
on maximum number of administration routes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{route$CTranspose()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-route$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/route.txt", "route")
obj$select()
obj$CTranspose()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-route-clone"></a>}}
\if{latex}{\out{\hypertarget{method-route-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{route$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
