% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPDextractor.R
\name{schedule}
\alias{schedule}
\title{schedule class}
\description{
R6 Class representing schedules (over-the-counter or prescription drugs).
}
\details{
' The package class has 2  methods in addition to 6 other methods and a field that inherits from the
parent class (DPDfile). The schedule table has 3 fields: "DRUG_CODE", "SCHEDULE", "SCHEDULE_F"
}
\examples{

## ------------------------------------------------
## Method `schedule$new`
## ------------------------------------------------

obj<-schedule$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/schedule.txt")


## ------------------------------------------------
## Method `schedule$select`
## ------------------------------------------------

obj<-schedule$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/schedule.txt")
obj$select()

}
\section{Super class}{
\code{\link[DPDextractor3:DPDfile]{DPDextractor3::DPDfile}} -> \code{schedule}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-schedule-new}{\code{schedule$new()}}
\item \href{#method-schedule-select}{\code{schedule$select()}}
\item \href{#method-schedule-clone}{\code{schedule$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DPDextractor3" data-topic="DPDfile" data-id="field_search"><a href='../../DPDextractor3/html/DPDfile.html#method-DPDfile-field_search'><code>DPDextractor3::DPDfile$field_search()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-schedule-new"></a>}}
\if{latex}{\out{\hypertarget{method-schedule-new}{}}}
\subsection{Method \code{new()}}{
Create a new schedule object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{schedule$new(fname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fname}}{file and path names.}

\item{\code{dpdlabel}}{distinctive element of the schedule name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new schedule object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-schedule$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/schedule.txt")

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-schedule-select"></a>}}
\if{latex}{\out{\hypertarget{method-schedule-select}{}}}
\subsection{Method \code{select()}}{
Chooses the table columns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{schedule$select(columns = c("DRUG_CODE"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{columns}}{field names.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-schedule$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/schedule.txt")
obj$select()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-schedule-clone"></a>}}
\if{latex}{\out{\hypertarget{method-schedule-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{schedule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
