% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPDextractor.R
\name{status}
\alias{status}
\title{status class}
\description{
R6 Class representing the DRUG product status. The status class has 5 methods in addition to 6 other methods and a field that inherits from the parent class (DPDfile).
}
\details{
A drug is considered to have 4 possible statuses: Approved, cancelled, marketed or dormant.
status.txt: Approved, marqueted ("Y")
status_ap.txt: Approved ("Y"), marqueted and cancelled
status_ia.txt: Approved, marqueted and cancelled  ("Y")
status_dr.txt: Approved, marqueted and dormant  ("Y")
}
\examples{

## ------------------------------------------------
## Method `status$new`
## ------------------------------------------------

obj<-status$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/status.txt", "status")
obj<-status$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/status.txt")

## ------------------------------------------------
## Method `status$Y_select`
## ------------------------------------------------

obj<-status$Y_select()

## ------------------------------------------------
## Method `status$select`
## ------------------------------------------------

obj<-status$select()
obj<-status$select(c("DRUG_CODE", "CURRENT_STATUS_FLAG", "STATUS"))

## ------------------------------------------------
## Method `status$DiffDate`
## ------------------------------------------------

obj<-status$DiffDate("30-JUN-2016")

}
\section{Super class}{
\code{\link[DPDextractor3:DPDfile]{DPDextractor3::DPDfile}} -> \code{status}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-status-new}{\code{status$new()}}
\item \href{#method-status-Y_select}{\code{status$Y_select()}}
\item \href{#method-status-select}{\code{status$select()}}
\item \href{#method-status-DiffDate}{\code{status$DiffDate()}}
\item \href{#method-status-clone}{\code{status$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DPDextractor3" data-topic="DPDfile" data-id="field_search"><a href='../../DPDextractor3/html/DPDfile.html#method-DPDfile-field_search'><code>DPDextractor3::DPDfile$field_search()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-status-new"></a>}}
\if{latex}{\out{\hypertarget{method-status-new}{}}}
\subsection{Method \code{new()}}{
Create a new status object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{status$new(fname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fname}}{file and path names.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new status object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-status$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/status.txt", "status")
obj<-status$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/status.txt")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-status-Y_select"></a>}}
\if{latex}{\out{\hypertarget{method-status-Y_select}{}}}
\subsection{Method \code{Y_select()}}{
Select current statuses
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{status$Y_select()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-status$Y_select()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-status-select"></a>}}
\if{latex}{\out{\hypertarget{method-status-select}{}}}
\subsection{Method \code{select()}}{
Choose the subset of the table columns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{status$select(columns = c("DRUG_CODE", "STATUS"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{columns}}{a vector of status column names}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-status$select()
obj<-status$select(c("DRUG_CODE", "CURRENT_STATUS_FLAG", "STATUS"))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-status-DiffDate"></a>}}
\if{latex}{\out{\hypertarget{method-status-DiffDate}{}}}
\subsection{Method \code{DiffDate()}}{
calculates how long ago a drug acquired a given status
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{status$DiffDate(fdate)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fdate}}{date on which a given drug acquired a status}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-status$DiffDate("30-JUN-2016")

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-status-clone"></a>}}
\if{latex}{\out{\hypertarget{method-status-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{status$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
