% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPDextractor.R
\name{ther}
\alias{ther}
\title{ther class}
\description{
R6 Class representing therapeutic class.
}
\details{
The package class has 2  methods in addition to 6 other methods and a field that inherits from the
parent class (DPDfile). The ther table has 4 fields:"DRUG_CODE", "TC_ATC_NUMBER", "TC_ATC" and "TC_AHFS_NUMBER"
(7 fields according to Readme section)
}
\examples{

## ------------------------------------------------
## Method `ther$new`
## ------------------------------------------------

obj<-ther$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/ther.txt")

## ------------------------------------------------
## Method `ther$select`
## ------------------------------------------------

obj<-ther$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/ther.txt")
obj$select()

}
\section{Super class}{
\code{\link[DPDextractor3:DPDfile]{DPDextractor3::DPDfile}} -> \code{ther}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ther-new}{\code{ther$new()}}
\item \href{#method-ther-select}{\code{ther$select()}}
\item \href{#method-ther-clone}{\code{ther$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DPDextractor3" data-topic="DPDfile" data-id="field_search"><a href='../../DPDextractor3/html/DPDfile.html#method-DPDfile-field_search'><code>DPDextractor3::DPDfile$field_search()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ther-new"></a>}}
\if{latex}{\out{\hypertarget{method-ther-new}{}}}
\subsection{Method \code{new()}}{
Create a new ther object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ther$new(fname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fname}}{file and path names.}

\item{\code{dpdlabel}}{distinctive element of the ther name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new ther object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-ther$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/ther.txt")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ther-select"></a>}}
\if{latex}{\out{\hypertarget{method-ther-select}{}}}
\subsection{Method \code{select()}}{
Chooses the table columns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ther$select(columns = c("DRUG_CODE", "TC_ATC"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{columns}}{field names.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-ther$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/ther.txt")
obj$select()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ther-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ther-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ther$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
