% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPDextractor.R
\name{vet}
\alias{vet}
\title{vet class}
\description{
R6 Class representing the species on which the drugs are used.
}
\details{
The package class has 4  methods in addition to 6 other methods and a field that inherits from the
parent class (DPDfile).
The vet table has 4 fields: "DRUG_CODE", "VET_SPECIES", "VET_SUB_SPECIES" and "VET_SPECIES_F"
}
\examples{

## ------------------------------------------------
## Method `vet$new`
## ------------------------------------------------

obj<-schedule$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/vet.txt")


## ------------------------------------------------
## Method `vet$select`
## ------------------------------------------------

obj<-vet$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/vet.txt")
obj$select()


## ------------------------------------------------
## Method `vet$CTranspose`
## ------------------------------------------------

obj<-vet$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/vet.txt")
obj$CTranspose()

}
\section{Super class}{
\code{\link[DPDextractor3:DPDfile]{DPDextractor3::DPDfile}} -> \code{vet}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-vet-new}{\code{vet$new()}}
\item \href{#method-vet-select}{\code{vet$select()}}
\item \href{#method-vet-CTranspose}{\code{vet$CTranspose()}}
\item \href{#method-vet-clone}{\code{vet$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DPDextractor3" data-topic="DPDfile" data-id="field_search"><a href='../../DPDextractor3/html/DPDfile.html#method-DPDfile-field_search'><code>DPDextractor3::DPDfile$field_search()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vet-new"></a>}}
\if{latex}{\out{\hypertarget{method-vet-new}{}}}
\subsection{Method \code{new()}}{
Create a new vet object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vet$new(fname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fname}}{file and path names.}

\item{\code{dpdlabel}}{distinctive element of the vet name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new vet object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-schedule$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/vet.txt")

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vet-select"></a>}}
\if{latex}{\out{\hypertarget{method-vet-select}{}}}
\subsection{Method \code{select()}}{
Chooses the table columns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vet$select(columns = c("DRUG_CODE", "VET_SPECIES", "VET_SUB_SPECIES"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{columns}}{field names.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-vet$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/vet.txt")
obj$select()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vet-CTranspose"></a>}}
\if{latex}{\out{\hypertarget{method-vet-CTranspose}{}}}
\subsection{Method \code{CTranspose()}}{
Selective transposition of drugs applied in at least one species
based on maximum number of species
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vet$CTranspose(ve = "VET_SPECIES")}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{obj<-vet$new("D:/WIAAgrosante/Mapaq-Antibioticos/DPD/vet.txt")
obj$CTranspose()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-vet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
